SQL_RAG/rag_app/backend/requirements.txt
# FastAPI backend requirements for SQL RAG application

# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Async HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Google Cloud dependencies
google-cloud-aiplatform==1.38.0
google-cloud-bigquery==3.13.0
google-cloud-firestore==2.13.1
google-auth==2.23.4

# Database and storage
pandas==2.1.4
numpy==1.25.2

# Configuration and utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
ruff==0.1.6